Correction bug + mise en conformité OCP

+++ BIancherin-Merle-Sustrac/src/Logger/Folder.java
@@ -5,19 +5,17 @@ import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 
-public class Folder {
+class Folder {
 
 	private long size;
 	private String Name;
 	private String Date;
-	private Message log;
 	private String path;
 	private BufferedWriter writer;
 	private File file;
 	private int i;
 
-	public Folder(long size, String Date) { // String path
-
+	protected Folder(long size, String Date) { 
 		this.size = size;
 		Name = "log" + i +"_"+GetDate()+ ".dat";
 		this.Date = Date;
@@ -30,62 +28,14 @@ public class Folder {
 		if(!(PropertiesFiles.propertiePresent("cible")&&
 				PropertiesFiles.displayOnePropertie("cible").equals("CONSOLE")))
 		{	
-			try {				
-				file = new File(path+Name);
-				writer = new BufferedWriter(new FileWriter(file, true));			
-				// writer.write(log);	
-				// writer.close();
-			} catch (IOException e) {
-				e.printStackTrace();
-			}
-		}
-	}
-
-	public long getSize() {
-
-		return size;
-	}
-
-	public void setSize(int size) {
-		this.size = size;
-	}
-
-	public String getName() {
-		return Name;
-	}
-
-	public void setName(String name) {
-		Name = name;
-	}
-
-	public String getDate() {
-		return Date;
-	}
-
-	public void setDate(String date) {
-		Date = date;
-	}
-
-	public void UpdateFolder(Message log) {
-				
-		//recuperation de la cible des logs dans le fichier properties
-		String target="ROTATE";
-		if(PropertiesFiles.propertiePresent("cible"))
-		{target = PropertiesFiles.displayOnePropertie("cible");}
-		if (file.length() > this.size & file.exists() & target.equals("ROTATE")) {
-			System.out.println("Size exceeded! New file created");
-			i = i + 1;
-			setName("log" + i+"_"+GetDate()+".dat");
-			
-
 			try {
-				file = new File(path+Name);
+				File filePath = new File(path);
 				//crée un sous dossier contenu dans le path si non existant
-				if(!file.exists())
+				if(!filePath.exists())
 				{
 					boolean result = false;
 					try{
-						file.mkdir();
+						filePath.mkdir();
 						result=true;
 					}
 					catch(SecurityException se){
@@ -94,18 +44,58 @@ public class Folder {
 					if(!result)
 					{System.out.println("Can't create the DIR");}
 				}
+				file = new File(path+Name);
+				writer = new BufferedWriter(new FileWriter(file, true));			
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+		}
+	}
+
+	protected long getSize() {
+		return size;
+	}
+
+	protected void setSize(int size) {
+		this.size = size;
+	}
+
+	protected String getName() {
+		return Name;
+	}
+
+	protected void setName(String name) {
+		Name = name;
+	}
+
+	protected String getDate() {
+		return Date;
+	}
+
+	protected void setDate(String date) {
+		Date = date;
+	}
+
+	protected void UpdateFolder(Message log) {			
+		//recuperation de la cible des logs dans le fichier properties
+		String target="ROTATE";
+		if(PropertiesFiles.propertiePresent("cible"))
+		{target = PropertiesFiles.displayOnePropertie("cible");}
+		if (file.length() > this.size & file.exists() & target.equals("ROTATE")) {
+			System.out.println("Size exceeded! New file created");
+			i = i + 1;
+			setName("log" + i+"_"+GetDate()+".dat");
+			try {
+				file = new File(path+Name);
 				writer = new BufferedWriter(new FileWriter(file));
 				// normalement si le fichier n'existe pas, il est crée à la
 				// racine du projet
-
 				writer.write("" + log);
 				writer.newLine();
-
 				writer.close();
 			} catch (IOException e) {
 				e.printStackTrace();
 			}
-
 		}
 
 		try {
@@ -115,19 +105,17 @@ public class Folder {
 			writer.newLine();
 			writer.write("" + log);
 			writer.newLine();
-
 			writer.close();
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
-
 	}
 
-	public String getPath() {
+	protected String getPath() {
 		return path;
 	}
 
-	public void setPath(String path) {
+	protected void setPath(String path) {
 		this.path = path;
 	}
 
@@ -137,7 +125,6 @@ public class Folder {
 				format);
 		java.util.Date date = new java.util.Date();
 		System.out.println(formater.format(date));
-
 		return formater.format(date);
 	}
